requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
Markdown==3.6
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
zipp==3.18.1
